+++ documentation/code/LineChartDemo.tsx
@@ -1,7 +1,11 @@
 import React from 'react';
 
 // eslint-disable-next-line shopify/strict-component-boundaries
-import {LineChart} from '../../src/components';
+import {
+  LineChart,
+  LineChartProps,
+  LineChartTooltipContent,
+} from '../../src/components';
 
 import {OUTER_CONTAINER_STYLE} from './constants';
 
@@ -17,82 +21,112 @@ export function LineChartDemo() {
   document.body.style.fontFamily =
     "-apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif";
 
-  const formatYAxisLabel = new Intl.NumberFormat('en', {
-    style: 'currency',
-    currency: 'CAD',
-    currencyDisplay: 'narrowSymbol',
-  }).format;
-  const formatY = new Intl.NumberFormat('en', {
-    style: 'currency',
-    currency: 'CAD',
-  }).format;
-
   const series = [
     {
       name: 'Apr 01–Apr 14, 2020',
       data: [
-        {rawValue: 2251, label: 'April 01, 2020'},
-        {rawValue: 12132.2, label: 'April 02, 2020'},
-        {rawValue: 5000, label: 'April 03, 2020'},
-        {rawValue: 7200, label: 'April 04, 2020'},
-        {rawValue: 1500, label: 'April 05, 2020'},
-        {rawValue: 6132, label: 'April 06, 2020'},
-        {rawValue: 3100, label: 'April 07, 2020'},
-        {rawValue: 2200, label: 'April 08, 2020'},
-        {rawValue: 5103, label: 'April 09, 2020'},
-        {rawValue: 2112.5, label: 'April 10, 2020'},
-        {rawValue: 4004, label: 'April 11, 2020'},
-        {rawValue: 6000, label: 'April 12, 2020'},
-        {rawValue: 5500, label: 'April 13, 2020'},
-        {rawValue: 7000, label: 'April 14, 2020'},
+        {rawValue: 2251, label: '2020-04-01T12:00:00'},
+        {rawValue: 12132.2, label: '2020-04-02T12:00:00'},
+        {rawValue: 5000, label: '2020-04-03T12:00:00'},
+        {rawValue: 7200, label: '2020-04-04T12:00:00'},
+        {rawValue: 1500, label: '2020-04-05T12:00:00'},
+        {rawValue: 6132, label: '2020-04-06T12:00:00'},
+        {rawValue: 3100, label: '2020-04-07T12:00:00'},
+        {rawValue: 2200, label: '2020-04-08T12:00:00'},
+        {rawValue: 5103, label: '2020-04-09T12:00:00'},
+        {rawValue: 2112.5, label: '2020-04-10T12:00:00'},
+        {rawValue: 4004, label: '2020-04-11T12:00:00'},
+        {rawValue: 6000, label: '2020-04-12T12:00:00'},
+        {rawValue: 5500, label: '2020-04-13T12:00:00'},
+        {rawValue: 7000, label: '2020-04-14T12:00:00'},
       ],
       style: {
         color: 'primary',
       },
-      formatY,
     },
     {
       name: 'Mar 01–Mar 14, 2020',
       data: [
-        {rawValue: 5200, label: 'March 01, 2020'},
-        {rawValue: 7000, label: 'March 02, 2020'},
-        {rawValue: 1000, label: 'March 03, 2020'},
-        {rawValue: 2000, label: 'March 04, 2020'},
-        {rawValue: 5000, label: 'March 05, 2020'},
-        {rawValue: 1000, label: 'March 06, 2020'},
-        {rawValue: 2000, label: 'March 07, 2020'},
-        {rawValue: 5000, label: 'March 08, 2020'},
-        {rawValue: 4000, label: 'March 09, 2020'},
-        {rawValue: 11200, label: 'March 10, 2020'},
-        {rawValue: 2000, label: 'March 11, 2020'},
-        {rawValue: 3000, label: 'March 12, 2020'},
-        {rawValue: 2000, label: 'March 13, 2020'},
-        {rawValue: 3000, label: 'March 14, 2020'},
+        {rawValue: 5200, label: '2020-03-01T12:00:00'},
+        {rawValue: 7000, label: '2020-03-02T12:00:00'},
+        {rawValue: 1000, label: '2020-03-03T12:00:00'},
+        {rawValue: 2000, label: '2020-03-04T12:00:00'},
+        {rawValue: 5000, label: '2020-03-05T12:00:00'},
+        {rawValue: 1000, label: '2020-03-06T12:00:00'},
+        {rawValue: 2000, label: '2020-03-07T12:00:00'},
+        {rawValue: 5000, label: '2020-03-08T12:00:00'},
+        {rawValue: 4000, label: '2020-03-09T12:00:00'},
+        {rawValue: 11200, label: '2020-03-10T12:00:00'},
+        {rawValue: 2000, label: '2020-03-11T12:00:00'},
+        {rawValue: 3000, label: '2020-03-12T12:00:00'},
+        {rawValue: 2000, label: '2020-03-13T12:00:00'},
+        {rawValue: 3000, label: '2020-03-14T12:00:00'},
       ],
       style: {
         color: 'pastComparison',
         lineStyle: 'dashed' as 'dashed',
       },
-      formatY,
     },
   ];
 
   const xAxisLabels = series[0].data.map(({label}) => label);
 
-  const formatXAxisLabel = (value: string) => {
-    const date = new Date(value);
-    return date.toLocaleDateString('en-CA', {day: 'numeric', month: 'numeric'});
+  function formatXAxisLabel(value: string) {
+    return new Date(value).toLocaleDateString('en-CA', {
+      day: 'numeric',
+      month: 'numeric',
+    });
+  }
+
+  function formatYAxisLabel(value: number) {
+    return new Intl.NumberFormat('en', {
+      style: 'currency',
+      currency: 'CAD',
+      currencyDisplay: 'symbol',
+      maximumSignificantDigits: 1,
+    }).format(value);
+  }
+
+  const renderTooltipContent: LineChartProps['renderTooltipContent'] = ({
+    data,
+  }) => {
+    function formatTooltipValue(value: number) {
+      return new Intl.NumberFormat('en', {
+        style: 'currency',
+        currency: 'CAD',
+      }).format(value);
+    }
+
+    function formatTooltipLabel(value: string) {
+      return new Date(value).toLocaleDateString('en-CA', {
+        day: 'numeric',
+        month: 'long',
+        year: 'numeric',
+      });
+    }
+
+    const formattedData = data.map(({name, point: {label, value}, style}) => ({
+      name,
+      style,
+      point: {
+        value: formatTooltipValue(value),
+        label: formatTooltipLabel(label),
+      },
+    }));
+
+    return <LineChartTooltipContent data={formattedData} />;
   };
 
   return (
     <div style={OUTER_CONTAINER_STYLE}>
       <div style={innerContainerStyle}>
         <LineChart
+          series={series}
           chartHeight={229}
           xAxisLabels={xAxisLabels}
           formatXAxisLabel={formatXAxisLabel}
           formatYAxisLabel={formatYAxisLabel}
-          series={series}
+          renderTooltipContent={renderTooltipContent}
         />
       </div>
     </div>