+++ documentation/code/StackedAreaChartDemo.tsx
@@ -1,20 +1,15 @@
 import React from 'react';
 
 // eslint-disable-next-line shopify/strict-component-boundaries
-import {StackedAreaChart} from '../../src/components';
+import {
+  StackedAreaChart,
+  StackedAreaChartProps,
+  TooltipContent,
+} from '../../src/components';
 
 import {OUTER_CONTAINER_STYLE} from './constants';
 
 export function StackedAreaChartDemo() {
-  const formatYAxisLabel = (value?: number) => {
-    const formatter = new Intl.NumberFormat('en').format;
-    if (value == null) {
-      return '-';
-    }
-
-    return formatter(value);
-  };
-
   const innerContainerStyle = {
     width: '900px',
     height: '300px',
@@ -39,21 +34,58 @@ export function StackedAreaChartDemo() {
       color: 'secondary',
     },
   ];
+
+  const labels = [
+    'January',
+    'February',
+    'March',
+    'April',
+    'May',
+    'June',
+    'July',
+  ];
+
+  const formatYAxisLabel = (value?: number) => {
+    const formatter = new Intl.NumberFormat('en').format;
+    if (value == null) {
+      return '-';
+    }
+
+    return formatter(value);
+  };
+
+  const renderTooltipContent: StackedAreaChartProps['renderTooltipContent'] = ({
+    data,
+    title,
+  }) => {
+    const formatTooltipValue = (val: number) =>
+      new Intl.NumberFormat('en').format(val);
+
+    const formattedData = data.map(({label, value, color}) => ({
+      color,
+      label,
+      value: formatTooltipValue(value),
+    }));
+
+    const total = data.reduce((totalValue, {value}) => totalValue + value, 0);
+
+    return (
+      <TooltipContent
+        title={title}
+        data={formattedData}
+        total={{label: 'Total', value: formatTooltipValue(total)}}
+      />
+    );
+  };
+
   return (
     <div style={OUTER_CONTAINER_STYLE}>
       <div style={innerContainerStyle}>
         <StackedAreaChart
-          formatYAxisLabel={formatYAxisLabel}
-          xAxisLabels={[
-            'January',
-            'February',
-            'March',
-            'April',
-            'May',
-            'June',
-            'July',
-          ]}
           series={series}
+          xAxisLabels={labels}
+          formatYAxisLabel={formatYAxisLabel}
+          renderTooltipContent={renderTooltipContent}
         />
       </div>
     </div>