+++ src/components/BarChart/Chart.tsx
@@ -1,4 +1,4 @@
-import React, {useState} from 'react';
+import React, {useState, useMemo} from 'react';
 import {Color} from 'types';
 
 import {eventPoint, getTextWidth} from '../../utilities';
@@ -6,8 +6,8 @@ import {YAxis} from '../YAxis';
 import {TooltipContainer} from '../TooltipContainer';
 import {StringLabelFormatter, NumberLabelFormatter} from '../../types';
 
-import {BarData} from './types';
-import {XAxis, Bar, Tooltip} from './components';
+import {BarData, RenderTooltipContentData} from './types';
+import {XAxis, Bar} from './components';
 import {useYScale, useXScale} from './hooks';
 import {
   MARGIN,
@@ -29,6 +29,7 @@ interface Props {
   formatXAxisLabel: StringLabelFormatter;
   formatYAxisLabel: NumberLabelFormatter;
   timeSeries: boolean;
+  renderTooltipContent: (data: RenderTooltipContentData) => React.ReactNode;
 }
 
 export function Chart({
@@ -40,6 +41,7 @@ export function Chart({
   formatXAxisLabel,
   formatYAxisLabel,
   timeSeries,
+  renderTooltipContent,
 }: Props) {
   const [activeBar, setActiveBar] = useState<number | null>(null);
   const [tooltipPosition, setTooltipPosition] = useState<{
@@ -49,7 +51,10 @@ export function Chart({
 
   const yAxisLabelWidth = data
     .map(({rawValue}) =>
-      getTextWidth({text: formatYAxisLabel(rawValue), fontSize: FONT_SIZE}),
+      getTextWidth({
+        text: formatYAxisLabel(rawValue),
+        fontSize: FONT_SIZE,
+      }),
     )
     .reduce((acc, currentValue) => Math.max(acc, currentValue));
 
@@ -86,6 +91,17 @@ export function Chart({
     data,
     formatYAxisLabel,
   });
+
+  const tooltipMarkup = useMemo(() => {
+    if (activeBar == null) {
+      return null;
+    }
+
+    return renderTooltipContent({
+      label: data[activeBar].label,
+      value: data[activeBar].rawValue,
+    });
+  }, [activeBar, data, renderTooltipContent]);
 
   return (
     <div
@@ -150,10 +166,7 @@ export function Chart({
           margin={MARGIN}
           position="center"
         >
-          <Tooltip
-            label={data[activeBar].label}
-            value={formatYAxisLabel(data[activeBar].rawValue)}
-          />
+          {tooltipMarkup}
         </TooltipContainer>
       ) : null}
     </div>