+++ src/components/LineChart/index.ts
@@ -1,1 +1,5 @@
-export {LineChart} from './LineChart';
+export {LineChart, LineChartProps} from './LineChart';
+export {
+  TooltipContent as LineChartTooltipContent,
+  TooltipContentProps as LineChartTooltipContentProps,
+} from './components';