+++ src/components/StackedAreaChart/Chart.tsx
@@ -1,7 +1,7 @@
 import React, {useState, useMemo} from 'react';
 import {stack, stackOffsetNone, stackOrderReverse} from 'd3-shape';
 
-import {TooltipContainer, Tooltip} from '../../components';
+import {TooltipContainer} from '../TooltipContainer';
 import {eventPoint} from '../../utilities';
 import {YAxis} from '../YAxis';
 import {Crosshair} from '../Crosshair';
@@ -13,15 +13,15 @@ import {Margin, Spacing} from './constants';
 import {useXScale, useYScale} from './hooks';
 import {StackedAreas} from './components';
 import styles from './Chart.scss';
-import {Series} from './types';
+import {Series, RenderTooltipContentData} from './types';
 
 interface Props {
   xAxisLabels: string[];
   series: Series[];
   formatXAxisLabel: StringLabelFormatter;
   formatYAxisLabel: NumberLabelFormatter;
+  renderTooltipContent(data: RenderTooltipContentData): React.ReactNode;
   dimensions: DOMRect;
-  tooltipSumDescriptor?: string;
   opacity: number;
   isAnimated: boolean;
 }
@@ -32,7 +32,7 @@ export function Chart({
   dimensions,
   formatXAxisLabel,
   formatYAxisLabel,
-  tooltipSumDescriptor,
+  renderTooltipContent,
   opacity,
   isAnimated,
 }: Props) {
@@ -95,35 +95,26 @@ export function Chart({
       return null;
     }
 
-    const labels = Object.keys(formattedData[activePointIndex]);
-    const values = Object.values<number>(formattedData[activePointIndex]);
-    const formattedValues = values.map(formatYAxisLabel);
-
-    const total = values.reduce((totalValue, value) => totalValue + value);
-    const formattedTotal = formatYAxisLabel(total);
-
-    return (
-      <Tooltip
-        colors={colors}
-        labels={labels}
-        values={formattedValues}
-        total={
-          tooltipSumDescriptor != null
-            ? {
-                label: tooltipSumDescriptor,
-                value: formattedTotal,
-              }
-            : null
+    const data = series.reduce<RenderTooltipContentData['data']>(
+      function removeNullsAndFormatData(tooltipData, {color, label, data}) {
+        const value = data[activePointIndex];
+        if (value == null) {
+          return tooltipData;
         }
-      />
+
+        tooltipData.push({color, label, value});
+        return tooltipData;
+      },
+      [],
     );
-  }, [
-    activePointIndex,
-    colors,
-    formatYAxisLabel,
-    formattedData,
-    tooltipSumDescriptor,
-  ]);
+
+    const title = xAxisLabels[activePointIndex];
+
+    return renderTooltipContent({
+      data,
+      title,
+    });
+  }, [activePointIndex, series, xAxisLabels, renderTooltipContent]);
 
   if (xScale == null || drawableWidth == null || axisMargin == null) {
     return null;