+++ src/components/TooltipContent/TooltipContent.tsx
@@ -3,34 +3,38 @@ import {Color} from 'types';
 
 import {SquareColorPreview} from '../SquareColorPreview';
 
-import styles from './Tooltip.scss';
+import styles from './TooltipContent.scss';
 
-interface Props {
-  colors: Color[];
-  labels: string[];
-  values: string[];
+interface TooltipData {
+  color: Color;
+  label: string;
+  value: string;
+}
+
+export interface TooltipContentProps {
   title?: string;
+  data: TooltipData[];
   total?: null | {
     label: string;
     value: string;
   };
 }
 
-export function Tooltip({title, colors, labels, values, total}: Props) {
+export function TooltipContent({title, data, total}: TooltipContentProps) {
   return (
     <div className={styles.Container}>
       {title == null ? null : <div className={styles.Title}>{title}</div>}
-      {labels.map((label, index) => (
+      {data.map(({color, label, value}, index) => (
         <React.Fragment key={`${label}-${index}`}>
-          <SquareColorPreview color={colors[index]} />
-          <p className={styles.SeriesName}>{label}</p>
-          <p className={styles.Value}>{values[index]}</p>
+          <SquareColorPreview color={color} />
+          <p className={styles.Label}>{label}</p>
+          <p className={styles.Value}>{value}</p>
         </React.Fragment>
       ))}
       {total == null ? null : (
         <React.Fragment>
           <div />
-          <p className={styles.SeriesName}>{total.label}</p>
+          <p className={styles.Label}>{total.label}</p>
           <p className={styles.Value}>{total.value}</p>
         </React.Fragment>
       )}